pyproject.toml
[tool.poetry]
name = "monitor"
version = "0.1.0"
description = ""
authors = ["Szymon Lopaciuk <szymon@lopaciuk.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^9.4.0"
requests = "^2.28.2"
python-dateutil = "^2.8.2"
click = "^8.1.3"
cairosvg = "^2.7.0"
importlib-resources = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
monitor = 'monitor.cli:entrypoint'